+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### improvements
 
+- Halt if Heroku app already exists.
 - Do not say "Pushing application to Heroku" if there is no deploy
 - Do not install Sendgrid unless user explicitely says so
 - Using user input for Heroku app name